unconditionally call PyClassLoader_UpdateSlot in tp_setattro
Summary:
We want to ensure that there is a consistent entry point for the python classloader - making sure that UpdateSlot always gets called allows us to place more checks in UpdateSlot, which will be important in the future as we want to reduce our specific hooks in random Objects/ files.

As part of the change, we also need to check for the vtable being null on the classloader side.

Since PyType_InitTypeForPatching gets called beforehand, we know that we will get vtables for static types. For non-static classes extending static types, we'll need to potentially even do work if vtable == NULL.

Reviewed By: carljm

Differential Revision: D33024791

fbshipit-source-id: f241ff5f7b