Bypass STORE_ATTR/DELETE_ATTR when setting up cached properties
Summary:
Currently, we use STORE_ATTR/DELETE_ATTR when a cached property
is set on a class. A side effect of this is, we end up calling
`_PyClassLoader_InitTypeForPatching`. This also initializes the vtable,
which is not a side-effect we want, because then we end up not generating
a slotmap for the cached property itself.

Reviewed By: carljm

Differential Revision: D33008020

fbshipit-source-id: e6a4668363