coerce int if CAST to float
Summary:
D33002386 (https://github.com/facebookincubator/cinder/commit/24012cbde550eb2fdb2734832be0b7ba47177af7) introduced an explicit CAST to float before trying to unbox to a double.
We follow PEP 484 in treating `int` as a subtype of `float` in type checking (or rather,
we fake it by treating the name `float` as meaning `int | float` in annotations), but this
means that code accepting a `float` and then trying to unbox it to double may fail
at runtime if it gets an int, even though it passed static checking.

Fix this by having `CAST` coerce int to float. This adds no overhead to the hot path,
since we only check for this case if the type check fails and we'd otherwise be
raising an exception.

Reviewed By: DinoV

Differential Revision: D33244884

fbshipit-source-id: b30ef5f5a0