Add test case for github issue #55 and fix issue in JITed code
Summary: This already seems to be doing the right thing in the interpreter, but it exposes a bug in JITed code.  When we have a primitive return type when we call into our helper we're not using the version that sets the error code in rdx.  This updates it so we have a thin wrapper to do that, as well as one for floating point.

Reviewed By: tekknolagi

Differential Revision: D33242324

fbshipit-source-id: c7856b23e8