Make property callable
Summary:
We currently don't treat `property` as being something that can be directly called.  This is just because we're missing `object` as the base type on it and we also don't reflect on it to pick up any ctor-like methods.

This updates `property` and fixes several other types missing `object` as a base.

Reviewed By: carljm

Differential Revision: D33244302

fbshipit-source-id: ffaa40ebe1