add support for pow for primitive doubles/ints
Summary:
Mostly for completeness - not really used in practice at the moment, but might come in handy.

One issue that I ran into is the fact that we assume the output type of a binop is a type we can always coerce the inputs to. This doesn't make sense for pow (2 ** -1 is 0.5), so let's always return doubles from these, and have the user explicitly choose to lose precision.

With regards to pow computation for explicitly signed unsigned + signed operands, I chose to disallow them as the implementation complexity seemed too much for the marginal benefit.

I also took the liberty to move some binop tests away from tests.py into their own file, which makes the diff look scarier than it is.

Reviewed By: carljm

Differential Revision: D33336332

fbshipit-source-id: 5e8b4308ee