don't infer literal from binop with literal operand
Summary:
Because we returned `self` or `other` from matched-signs cases in
`validate_mixed_math`, we would sometimes wrongly infer that a literal
type from one of the operands would carry through to the result of the
binop. Fix by never returning `self` or `other` directly, instead index into
the canonical signed/unsigned types lists.

There's an additional remaining potential bug here, in that some places
we check e.g. `typ in SIGNED_CINT_TYPES` to see if a type is a signed
`CIntType`, but this will wrongly exclude literal types from the check.
Added a TODO to follow up in separate diff.

Reviewed By: DinoV

Differential Revision: D33478925

fbshipit-source-id: fd7198698d