eliminate emit_augassign method
Summary:
The `emit_augassign` method on `Value` is not overridden by any type
and always call back to the default `emitAugAssign`, so let's simplify the
compiler by eliminating it.

I'd like to eliminate augassign-specific logic from the static types entirely;
it should all boil down to more generic load/store operations on the types
and the details of aug-assignment can be handled in just one place in the
code generator.

Reviewed By: czardoz

Differential Revision: D33483136

fbshipit-source-id: 141ef037e3