Fix memory leak in posix module
Summary:
D32571983 adds a C++ test that tears down and reinits the Python
runtime within a single test, which exposed the leak of `billion` in
`posixmodule.c`. There's probably some cleanup that can be done to avoid a
static variable like this, but the simplest fix for now is to add a free
function for the module, releasing the reference to `billion` there.

I also cleaned up our ASAN testing situation a little: `ASAN_TEST_ENV` includes
`detect_leaks=0`, from D23639621. I did a quick run of `make testcinder` with
it removed, and there are still a bunch of leaks that I don't want to get
sidetracked by now. `runtime_tests` is easier to keep leak-free, though, so I
added a separate variable to use with LSAN enabled, `ASAN_TEST_ENV_WITH_LSAN`.

Reviewed By: jbower-fb

Differential Revision: D33463323

fbshipit-source-id: 9ca1def549