get richards_static working via static loader
Summary:
Extract the bulk of `richards_static.py` into a `richards_static_lib.py` with
`import __static__`, and make `richards_static.py` a thin wrapper that
just imports and runs the lib.

Now that Static Python modules are compiled to rely on the existence
of a `<fixed-modules>` in their namespace, it's no longer possible to
compile a static pyc and run it directly as a main script via
`./python mystaticmod.pyc`. A static script can still be run using
`./python -m compiler --static ...`, but this does not use module
loaders and thus does not use pyc files, so every run will recompile
the module. This makes it hard to get a fair overall time comparison
with non-static.

After this diff, static Richards can be run with just
`python -X install-strict-loader richards_static.py`, which is
a direct analogue to `./python richards_static.py` for the non-static
comparison.

Also make it possible to pass the number of Richards iterations in via
CLI argument.

Reviewed By: DinoV

Differential Revision: D33519822

fbshipit-source-id: 1320dca9bf