fix slice subscript of array
Summary:
We had a static compiler test and a runtime non-static test for array
slicing, but no test executing static compiled bytecode. This allowed
a bug to slip through where we wrongly tried to box/unbox the value
going to/from a sliced load/store, assuming it must be the contained
type, whereas for a slice it's actually another array. Because
`REFINE_TYPE` and boxing/unboxing are basically no-ops in the
interpreter loop, the resulting bad bytecode showed up as crashes
in jit compilation.

Reviewed By: czardoz

Differential Revision: D33522013

fbshipit-source-id: 823bb7903e