don't emit unnecessary CONVERT_PRIMITIVE
Summary:
Literal type objects don't compare equal to the base primitive type
of the same size, which caused us to emit useless
`CONVERT_PRIMITIVE` from and to the same size in primitive binops
involving a literal as one operand. Prevent this by checking the actual
size opargs we are about to emit against each other.

Reviewed By: czardoz

Differential Revision: D33522697

fbshipit-source-id: 090d89444a