add static python version of fannkuch
Summary:
Static Python version of fannkuch benchmark.

This still does the same work as the original, but it's significantly
modified to better take advantage of Static Python. It uses arrays
instead of lists, and replaces complex list slicing and inserts/pops with
simple while loops and individual element indexing. We can't do slicing
or call insert/pop methods with primitives, so the original version
requires lots of boxing/unboxing if we try to use primitives, whereas
the updated version can use `SEQUENCE_GET`/`SEQUENCE_SET`
and work entirely with primitive ints.

The updated version is _much_ slower in non-static Python, because
it moves work out of optimized C list methods and into Python code.
But it's much faster in Static Python. In a longer-term sense, this is
a clear improvement, because it suggests that in the future,
list methods that are currently implemented in C could be implemented
in Static Python instead, with improved performance.

Reviewed By: DinoV

Differential Revision: D33531594

fbshipit-source-id: ce050c3210