support unary not on cbool/cint
Summary:
This support was missing and I wanted it for static Richards.

Also noticed as part of implementing this that we didn't have
support in place for implicitly boxing Python bool to cbool in
nonstatic -> static calls. This lack of support can easily pass
silently since we'll just fall back to deopting and vectorcalling
the function in the interpreter loop. Added support for cbool args
for now but we'll want to revisit this to find a more thorough way
to verify we support this for all primitive types; it looks like double
is also missing here.

Reviewed By: sinancepel

Differential Revision: D33533799

fbshipit-source-id: b96f487324