Walk the shadow frame stack in GDB
Summary:
Currently, if we crash in the JIT, there's no easy way to
extract the Python stacktrace which led to the crash. We can get hints
with the `py-walk-frame` command, but that's unreliable because it
relies on some guesswork.

In this diff, I've added two new GDB commands, `py-walk-shadow-sync` and
`py-walk-shadow-async` which walk the shadow frame stack. Together,
these will let us root cause issues much faster.

Reviewed By: akatrevorjay

Differential Revision: D33473380

fbshipit-source-id: d68b989e87