fix array fast for loop
Summary:
In D33530295 (https://github.com/facebookincubator/cinder/commit/97a0260fdc6b356e078b6d1c077c39aeef159b05) I forgot to add the new array test file to `__init__.py`, so it
didn't run in CI, and I missed that `test_fast_forloop` failed in the JIT.

The failure was due to emitting `LOAD_FAST` and `STORE_FAST` for
the loop target variable, when for an array with primitive members it
should be `LOAD_LOCAL` and `STORE_LOCAL`.

Root cause of this was that `ArrayInstance` didn't implement
`get_iter_type`, so type binding didn't bind the correct type to the
loop target.

Fix the bug and add the array tests so they'll run in CI.

Reviewed By: wmeehan

Differential Revision: D33550651

fbshipit-source-id: 10f9df824c