Add support for profile data version 3, with dict keys
Summary:
This adds a new profile data version along with runtime support for
eagerly initializing the cached dict keys of any types found in the given
profile data. This is necessary to support LOAD_ATTR when the receiver has
split dict attrs; we need the cached keys on the type to be in their final
state before JIT compilation happens in order to perform the specialization.

It also contains a small fix to ensure that `_PyJIT_TypeCreated()` is called as
late as possible. `type_new()` does some work after calling `PyType_Ready()`,
so don't call `_PyJIT_TypeCreated()` for heap types from `PyType_Ready()`,
instead calling it at the end of `type_new()`.

Reviewed By: carljm

Differential Revision: D36839855

fbshipit-source-id: 138baa423510d2515dd2c6c19203c4eb4abf9600