Make outputType take a lambda and fix GuardTypeRemoval
Summary:
In order for `GuardTypeRemoval` to properly handle passthrough
instructions, we need to be able to examine their output types if passed
in a different operand. This diff adds an `outputType` function that
takes in a lambda that can be used to get the instruction's operands'
types during the output calculation.

Once we're able to get the `outputType`s of passthrough instructions, we
just need to check whether the output type of the instruction given the
non-guarded type would still type-check successfully. In order to avoid
overly large call-stacks, we can do this iteratively by keeping a worklist
of passthrough output registers + types that still need to be checked.

We should also keep around the set of states we've already checked to
ensure that we don't enter an infinite loop when examining a `Phi`
instruction which uses its own output

Reviewed By: swtaarrs

Differential Revision: D33107329

fbshipit-source-id: b018af9bd9