Add simplifyCompare for Is/IsNot and between two Nones
Summary:
Introduce a few `simplifyCompare` cases based off of known types:

Since `Compare` `Is/IsNot` is based off of pointer equality, we can make a
few assumptions based off of the type of the arguments alone. If the two
arguments cannot be the same type, `Compare Is` can be turned into a
`LoadConst<False>` and `Compare IsNot` can be turned into a
`LoadConst<True>`. If both types are specialized, we can go a step further
and make the pointer comparison at compilation time, allowing us to emit
the `LoadConst` based off of the result of comparing the two `objectSpec`s.

Since `None`s are a singleton object, `Compare` of `Equal` of two `None`s
can be turned into `True` and `Compare` of `NotEqual` can be turned into
`False`.

By changing the `Compare` directly to a `LoadConst` instead of the
`PrimitiveCompare` that gets generated by `DynamicComparisonElimination`,
we allow its users to potentially further simplify at no extra cost.

Reviewed By: swtaarrs

Differential Revision: D32811447

fbshipit-source-id: 6cf134a566