Add hir::LongCompare to optimize long-only comparisons
Summary:
If we know that the two operands of a `COMPARE_OP` are `Long`s, we
should be able to emit a more specialized opcode that directly calls
`PyLong_Type.tp_richcompare`, and that has no memory side-effects.

Note: By turning `Compare` into `LongCompare` before running
`DynamicComparisonElimination`, we remove the ability to turn a `Compare`
and `IsTruthy` into a single `CompareBool`. However, the actual work done
by a `CompareBool` between two `Long`s should be the same as the work
done by `LongCompare` + `IsTruthy`.

Reviewed By: carljm

Differential Revision: D32998670

fbshipit-source-id: 6f2ae167f9