Introduce `typed_descriptor_with_default_value`
Summary:
One common pattern we don't support today is users wanting to declare instance variables for a class at the class toplevel, while giving it a default. This causes a lot of issues when converting modules to static, as we previously had to annotate all of these as ClassVar's and deal with the fallout.

This diff introduces a `typed_descriptor_with_default_value` descriptor which mimics the vanilla Python behavior more closely. The idea is that when initializing a class, we create slots for the `x: int = default` pattern which hold a reference to the default, and return the default if the instance doesn't have the slot filled out.

For this diff, we keep things simple and compile static loads/stores down to a `LOAD_ATTR` and `STORE_ATTR`. A future diff will add invoke support which should make things more efficient.

Reviewed By: czardoz

Differential Revision: D33407790

fbshipit-source-id: 14f539656b