fix multiple imports of same module
Summary:
If a module is compiled (via `load_compiled_module_from_source`) multiple times,
without first being imported from another static module, we would re-create a new
AST each time, thus causing KeyErrors in type binding on the second import
because it would bind a different AST than the one that had gone through decl visit.

The root cause of this problem is that our AST cache wasn't necessarily kept in
sync with our module dict; after `load_compiled_module_from_source` we'd have
an entry for the module in `self.modules` but no AST cache saved for it. Thus on
the second import we would not re-run decl visit (because the module table already
exists in `self.modules`) but we would re-run type binding with a new AST, causing
the KeyErrors.

Fix this by moving the AST cache down into the static Compiler and ensuring it is
kept in sync with `self.modules`, by always populating it in `add_module`.

Another indication this is the right approach is that it allows us to remove the
separate AST cache that was previously kept by our `TestCompiler`; the main
AST cache now covers that use case also.

Reviewed By: wmeehan

Differential Revision: D33660145

fbshipit-source-id: dd94ca1934