use sys.path as default strict analyzer import path
Summary:
We now have a default set of strict module stubs, but those make
heavy use of  `implicit`, and so a lot of them are pretty useless if
the standard library can't be found on the strict analyzer's import
path.

Let's complement the default stubs by also setting `sys.path` as the
default import path for the strict analyzer. This makes it more possible
in practice to use the strict/static loader via `-X install-strict-loader`,
where you can't set custom config on the loader.

Reviewed By: czardoz

Differential Revision: D33719499

fbshipit-source-id: 46ba255d72