Make `RaiseStatic` use a copy of the supplied `FrameState`
Summary:
`RaiseStatic` was constructing a fresh, empty `FrameState` and only
initializing its next bytecode offset. As a result, we would always
unwind immediately, rather than executing any `except` or `finally`
blocks that the instruction was contained within.

I ran into this when trying to fix the discrepancies in output between
`make testcinder` and `make testcinder_jit{_shadowframe}`. When run
under the JIT, the finally block in `test_double_await.runner` below
(from test_pep492.py) wasn't executed:

```
    def test_double_await(self):
        async def afunc():
            await asyncio.sleep(0.1)

        async def runner():
            coro = afunc()
            t = self.loop.create_task(coro)
            try:
                await asyncio.sleep(0)
                await coro
            finally:
                t.cancel()

        self.loop.set_debug(True)
        with self.assertRaises(
                RuntimeError,
                msg='coroutine is being awaited already'):

            self.loop.run_until_complete(runner())
```

As far as I can tell, this is only used when lowering `GET_AWAITABLE`
to raise a runtime error if the iterable is a coroutine and it's already
been awaited.

Reviewed By: tekknolagi

Differential Revision: D33720435

fbshipit-source-id: df6d15ae1e