Update `f_lasti` when materializing frames
Summary:
Update `f_lasti` when materializing `PyFrameObject`s for JIT-compiled
functions. This should product correct line numbers in warnings
and things that walk the Python stack.

We need to solve two problems to make this work. First, we need to
figure out which shadow frames need to have their associated
`PyFrameObject`s updated. Second, once we know which frames need to
be updated, we need to find the correct value for `f_lasti`.

We steal the last available tag bit in the `data` member of the
shadow frame struct to indicate which shadow frames should have
their `PyFrameObject`s updated. The bit stores who owns the shadow
frame - either the JIT or the interpreter. Shadow frames owned
by the JIT should have their associated `PyFrameObject`s updated.

To compute `f_lasti`, we construct a mapping of program points to the
bytecode instruction offset that produced them at compile-time. At runtime, we
read the saved return address for each JIT-compiled function
that is on the call stack. This is possible because 1) every function
has a fixed-size stack-frame and 2) the function's shadow-frame is
embedded at the very beginning of its stack-frame. We recover `f_lasti`
by looking up the saved return address in the mapping that we constructed
at compile-time.

It's worth noting that this approach only works if 1) calls are the
only way to leave a JIT-compiled function and enter another function
and 2) we instrument all calls for which we need to track bytecode
offset information. In practice I think that this approach works
pretty well. For 1, it should be rare to enter another function via something
other than a call. For 2, the majority of our codegen logic is centralized
in autogen so there's only a handful of places that need to be
instrumented.

Reviewed By: jbower-fb

Differential Revision: D33594435

fbshipit-source-id: c795187cd6