fix local types in loops
Summary:
Because we didn't do any kind of fixed-point iteration on loop bodies,
we weren't able to correctly handle cases where the local type of some
variable is widened within the loop body, and a use of it prior to that
widening relies on the widened type (see added tests.)

Add fixed-point iteration on loop bodies to handle these cases.

Because the fixed-point iteration requires comparing local type dicts
for equality, it surfaced the fact that literal types were created anew
every time, so one `Literal[1]` would be different from another and not
compare equal. Rather than implementing `__eq__` I decided to keep
identity of Value objects meaningful, and implemented a cache for
literal types in the compiler, parallel to the existing one for generic
types.

There are also a few small refactorings to how error sinks are handled,
in order to make it easier to temporarily replace the error sink on a
TypeBinder. (When fixed-point-iterating we need to use a
`CollectingErrorSink` and ignore errors.)

Reviewed By: sinancepel

Differential Revision: D33723824

fbshipit-source-id: d3625d4942