add low-touch typed DeltaBlue
Summary:
Add two low-touch conversions of DeltaBlue to Static Python.

`deltablue_static_basic.py` is a super-low-touch conversion; it is simply
the result of running MonkeyType to add annotations for the existing
types, and then adding a few casts for places where the existing code
relies for type safety on invariants that are beyond the current
Python type checkers, so the code will compile with Static Python.

`deltablue_static_basic2.py` is the same as the first, except using
`CheckedList` in place of the `OrderedCollection(list)` subclass used
by the prior code.

Reviewed By: DinoV

Differential Revision: D33747942

fbshipit-source-id: 3e8e90e332