Fix: Update line numbers used in getframe tests
Summary:
Another change landed that added code to `Lib/test/test_cinderjit.py`
before the tests that check for correct line numbers for frames
returned by `getframe`. This caused the line numbers used in the
tests to be incorrect.

Move the line number tests to the top of the file so that they're
less likely to be affected by additions. Additionally, refactor
the tests a bit so that they're easier to update if line numbers
do change.

Reviewed By: carljm

Differential Revision: D33772093

fbshipit-source-id: 37f0cd89b3