Fix JIT boxing for enum returns
Summary:
We were unconditionally boxing enums in the JIT upon function exit to non-JIT code.
However, static non-JIT code should receive a boxed integer for the eval loop.
Check whether the function was invoked statically and call the correct box function.

Change the jump for arg reentry to a long jump to allow for longer generic entry points.
This requires a change to the `JITRT_CALL_REENTRY_OFFSET` and `JITRT_STATIC_ENTRY_OFFSET`
to make more room for the long jump instruction.

Reviewed By: carljm

Differential Revision: D33108562

fbshipit-source-id: 829ee3d3bb