make type environment responsible for tracking + creating generic types
Summary:
We were previously breaking the abstraction barrier a bit by having the type objects be aware of the
instantiated generic type cache and referring to it themselves. Make the type environment be responsible for creating + tracking these generics.

There is one case that I'm not sure how to best resolve that's preventing us from avoiding passing in the type environment when instantiating a generic type - the GenericClass also makes all of its members generic. Previously, since we had direct access to the cache, the code to create a reference to the type we're instantiating + updating the members after to avoid infinite loops was sensible. With my changes, it looks awkward as we break the abstraction, which is an unforunate side effect of the change.

Reviewed By: carljm

Differential Revision: D33793080

fbshipit-source-id: b44f227663