add optimized Static Python DeltaBlue
Summary:
This is an optimized Static Python deltablue benchmark, using
primitives and a bit of inlining.

Main cost at this point is deopting in `get_planner()` because of the
dumb usage of a mutated global `planner` variable, but this would
require a significant rewrite to get rid of. Also some cost due to
having to box primitive arguments and return values in an
`INVOKE_METHOD`.

Reviewed By: wmeehan

Differential Revision: D33771577

fbshipit-source-id: 15cd9e29a3