Fix: Work around difference in refcounting behavior between the JIT and interpreter
Summary:
There's a small difference between the JIT and the interpreter around the
timing of when the last references to locals are destroyed in generators.
In the JIT, these are destroyed by refcounting ops inserted into the
compiled code. The references are destroyed before returning to the
gen_send implementation. In the interpreter, the references are held
by the generator's PyFrameObject. The generator's frame isn't destroyed
until after returning to the gen_send implementation.

This difference was causing a bug in some test code that assumed that a
finalizer wouldn't run while a generator that it referenced was also
running (see the comment in the code for full details).

Reviewed By: jbower-fb

Differential Revision: D33810627

fbshipit-source-id: 68197edff5