Use floating point registers
Summary:
This fixes a number of issues around our usage of floating point registers.  It's kind of several distinct fixes, but it's hard to break these out because when testing them they all end up becoming intertwined.

First, currently we setup the backend to expect only general purpose registers.  We need to flow information from the front end about what the inputs are so that we can properly interleave the usage of general purpose and floating point registers.  This now causes `arg_locations` to be dynamically populated.

Second, the static entry point is not currently preserving all of the registers it needs to.  When we call out to link the frame this results in the xmm registers being potentially clobbered because they are all caller save.  This also requires an increase in the size of the `jmp` used for the static entry point as we can end up saving a lot of registers.

Finally, autogen currently bases it's decision on whether to use the `X` format based upon the data type rather than if the register is actually an `xmm` register.  This leads to various bad code gen cases where we end up having a non-double typed xmm0 register, which can happen with things like non-initialized xmm registers which end up being `TNullptr` or `TCDouble.`  This should probably be cleaned up, but there appear to be multiple sources of this.  This cleans up at least one of those with `LoadArg` now properly getting the output type.

Reviewed By: carljm

Differential Revision: D33594087

fbshipit-source-id: 6569d95df8