Emit initialization of unassigned primitive values
Summary:
Currently primitive values are relying upon the JIT zero-initializing them.  This leads to some funky behavior in the JIT where we end up with `TNullptr | TC*` which turns our primitives into `TObject` in the back end.

It's rather difficult to fix the types that we initialize the values to as the zero-initialization occurs before we re-flow the types in SSA.  We could recognize all the combinations of the types in the backend, or we can perform the analysis in the compiler and zero initialize them.

This also opens the door to requiring the programmer to zero initialize the values instead of doing so automatically, but we probably need to make sure our existing usages of primitives are ready for that.

Note: the definite assignment visitor here is mostly just stolen from static Python, and updated to work on a per-method basis and just record the unvisited variables.

Reviewed By: carljm

Differential Revision: D33689976

fbshipit-source-id: 52a4342468