Add static version of nbody
Summary:
This adds a static version of `nbody`.  The main thing is switching from using tuples as data structures to using classes.  There's a new `Vector` class which stores an x, y, z triplet of doubles, as well as a new `Body` which stores two vectors and a double for mass.

There's then appropriate typing added throughout the benchmark.

Reviewed By: carljm

Differential Revision: D33770639

fbshipit-source-id: 75a0ba5af1