add force-strict hook to loader
Summary:
Allow a custom subclass of `StrictSourceFileLoader` to define its own
criteria for determining whether a to-be-loaded module should be
forced strict. Subclasses can override the new `should_force_strict`
method, which can access `self.path` or `self.name` or other loader
attributes to decide whether a module should be forced strict.

This also required a small change to the strict checker so that a
forced-strict module is not only analyzed as strict, but actually returns
marked as a strict module. This doesn't break any tests; I think it's
more reasonable behavior, just a case that was never relevant to our
previous (minimal) uses of forced-strict.

Reviewed By: czardoz

Differential Revision: D33929589

fbshipit-source-id: b8968403f6