Fix env var name for gdb-stubs-support
Summary:
This renames the enviroment varibale used to control gdb stub support to `PYTHONJITGDBSTUBSSUPPORT` previously it was `PYTHONJITGDBSUPPORT` which was used to control debug and gdb support as well. These should be seperate as different command line argument are used to control these different features (in the form of `jit-gdb-support` and `jit-gdb-stubs-support`).

I cannot find any code which uses `PYTHONJITGDB[STUBS]SUPPORT` so should be low risk change.

Reviewed By: carljm

Differential Revision: D33941065

fbshipit-source-id: 962de69a14