don't return forced-strict module as strict from analyzer
Summary:
In D33929589 (https://github.com/facebookincubator/cinder/commit/153b650de4fd445959e1f550ae4a43456f33403c) I changed the behavior of strict analyzer so if a module is
forced-strict the analyzer returns an AnalyzedModule with ModuleKind of
Strict. Turns out that change did break some assumptions in client code. On
second thought it's better not to make this change, as it means we lose any
way to find out from the analyzer whether the module was actually explicitly
marked as strict when we are forcing strict.

Instead of that change, the compiler can consider the module strict if it was
forced strict and came back with no errors.

In order to simplify this change, avoid returning an entire `StrictAnalysisResult`
from the compiler back to the loader, when all it will do with it is check the
module-kind; instead return just an `is_valid_strict` bool. And take advantage
of that to add some extra confidence in tests that the code we try to strict
compile actually is compiled as strict. (This last is pretty much unrelated, I
just noticed it and took advantage of the new simplified return type.)

Reviewed By: itamaro

Differential Revision: D33983046

fbshipit-source-id: 97ae1e0f8b