prevent unsound static->nonstatic->static MRO
Summary:
If a static type `C` inherits a nonstatic type `B`, `B` is opaque to the compiler,
so we don't know what it inherits.

If `B` inherits from another static type `A`, this leads to unsoundness,
because the compiler can't verify that the methods and attributes of `C`
are valid overrides of the methods and attributes of `A`, but instances of `C`
will pass runtime type-checks for `A`.

This can't be caught in the compiler (because `B` is opaque to the compiler),
so let's catch it when constructing `C` at runtime, in `set_type_static`.

Reviewed By: wmeehan

Differential Revision: D33279510

fbshipit-source-id: 44f20c2ef0