don't mutate classes during type binding
Summary:
ContextDecorator was mutating the containing class type during type binding.
This is not safe, since the implicit contract of our compiler (which needs to be
made more explicit) is that declaration visit results in a fully populated and
final module table, and it shouldn't make a difference to other modules
whether some other module has been type-bound or not; type binding should
be pure and should not change anything visible in the module table. If we
violate this contract, it can cause (and has caused) super-sneaky bugs where
the results of type binding module X can depend on whether or not module Y
has been type-bound yet: effectively, import side effects in our compiler!

It also caused a bug where importing (and thus type binding) a module twice
would throw an error on the second occurrence of this class mutation, since
it looked like trying to define the same slot twice on the class.

Fix this by expanding the existing `finish_bind` hook slightly, such that it takes
an optional `Class` (when we are finishing binding a class member), and updating
`Class.finish_bind` so that it can cope with new members being added to the class
by some other members' `finish_bind` (as `ContextDecorator` now does.)

This required also improving consistency of `finish_bind` handling a bit on functions
in general. `ModuleTable.finish_decorator` got inlined into `Function.finish_bind`
(it really should never be called independently) and we now make sure that when
we wrap a function in a decorator, we `finish_bind` the new decorator also, so
everything that ends up in a class' members or module's children has had
`finish_bind` called on it.

Reviewed By: wmeehan

Differential Revision: D33865534

fbshipit-source-id: 43ac7013ad