snap to original builtins dictionary
Summary:
Currently we're not capturing and using the starting builtins, which results in issues where something shows up in builtins later but is also imported in a module.  If the cached pyc is used before the value has been published in builtins then we'll use the cached version.

This causes us to always snap to the cached built in value.

Reviewed By: carljm

Differential Revision: D34018088

fbshipit-source-id: 306a6f989d