increase py_compile recursion limit
Summary:
The python compiler hits python recursion limit when running on the fbcode wheel.
Since the python AST walker is very recursion heavy, ~~I increased the recursion limit when the AST walker is created. Maybe there are better places to put the increase?~~
The recursion limit is now set in `compileall` when the default C compiler is not used

Reviewed By: carljm

Differential Revision: D33988587

fbshipit-source-id: 788cb2c9a2