Fix calls from non-static python to floating point functions
Summary:
Fixes up calls from non-static Python to static functions which have floating point arguments/return values.  Currently this is "working" because we go through the call with primitive arguments helper.  But we don't handle floating point values there, and end up falling back to the interpreter where we're supposed to report an error - but of course things just work in the interpreter.

Once that's fixed we now make it into the JITed code, where other issues are exposed.  We don't preserve the correct value around unlinking the frame, we don't use the right return type for the helper (although practically nothing clobbers xmm0/xxm1 here so we get away with it today), and we're not properly loading the values from the args array into the correct registers.

Reviewed By: carljm

Differential Revision: D33908637

fbshipit-source-id: af14bbd4fd