Skip unsupported distutils and grp tests
Summary:
From Matt's original commit for removing the negative test for non-existent group ids:
Unfortunately, on dev servers, `grp.getgrall()` returns only a subset
of gids that valid inputs to `grp.getgrgid()`. I'm not sure what sort
of shared library shenanigans we're playing on dev servers to make this
happen, but it seems unlikely that we will make a change that will break
this functionality.

Reviewed By: carljm

Differential Revision: D34062690

fbshipit-source-id: 1caf219335f2b92b3abae1a2c6592e42328b58b9