refactor all builtin types to be under a separate namespace
Summary:
This diff moves the scope of all builtin type construction to be within a class instead of being on the global namespace. The basic idea is that we want all builtin types to have their lifetime managed by the type environment so that we can support exact and readonly versions easily. This diff is a step in the way - we untangle the reliance on global state during construction and change things up so that the module-level computation is limited.

For now, we have a BuiltinTypes class which takes over the work that the different global types used to do - now that we want to avoid naming globals directly, we thread a reference to the builtin_types in the construction of classes. This diff also makes it so we don't introspect via `make_type_dict()` until all builtin type references are created. This allows us to avoid having a global dict for bootstrap purposes.

In future diffs, we will remove the BuiltinTypes class entirely and make things part of the type environment. For now, I keep the global BUILTIN_TYPES object in to minimize the changes we needed to make. The next immediate diff will remove BUILTIN_TYPES and scope things to be per-test and construct the builtin types when bringing up the compiler, but I wanted to keep the diff reviewable.

Reviewed By: wmeehan

Differential Revision: D34020575

fbshipit-source-id: 0862d281ac